refactor: share stripe utils
