+++ supabase/functions/create-checkout-session/index.ts
@@ -1,6 +1,6 @@
 import { serve } from "https://deno.land/std@0.168.0/http/server.ts";
 import Stripe from "https://esm.sh/stripe@13.2.0?target=deno";
-import { createClient } from "https://esm.sh/@supabase/supabase-js@2.38.0";
+import { authenticateUser, getOrCreateCustomer } from "../_shared/stripeUtils.ts";
 
 // Types
 interface CheckoutRequest {
@@ -58,65 +58,6 @@ const getPlanConfig = (planId: string, isYearly: boolean): PlanConfig => {
   }
 };
 
-// Authentication
-const authenticateUser = async (authHeader: string | null) => {
-  if (!authHeader) {
-    throw new Error("Autorização necessária");
-  }
-
-  const supabaseUrl = Deno.env.get("SUPABASE_URL");
-  const supabaseAnonKey = Deno.env.get("SUPABASE_ANON_KEY");
-
-  if (!supabaseUrl || !supabaseAnonKey) {
-    throw new Error("Variáveis do Supabase não configuradas");
-  }
-
-  const supabase = createClient(supabaseUrl, supabaseAnonKey);
-  const token = authHeader.replace("Bearer ", "");
-  const { data: { user }, error: authError } = await supabase.auth.getUser(token);
-
-  if (authError || !user) {
-    throw new Error("Usuário não autenticado");
-  }
-
-  console.log("✅ Usuário autenticado:", user.email);
-  return { user, supabase };
-};
-
-// Customer management
-const getOrCreateCustomer = async (
-  stripe: Stripe,
-  supabase: any,
-  user: any
-): Promise<string> => {
-  // Verificar se o usuário já tem um cliente Stripe
-  const { data: subscriber } = await supabase
-    .from("subscribers")
-    .select("stripe_customer_id")
-    .eq("user_id", user.id)
-    .single();
-  
-  let customerId = subscriber?.stripe_customer_id;
-
-  if (!customerId) {
-    console.log("🆕 Criando novo cliente Stripe");
-    const customer = await stripe.customers.create({
-      email: user.email,
-      metadata: {
-        user_id: user.id,
-      },
-    });
-    customerId = customer.id;
-    
-    await supabase.from("subscribers").upsert({
-      user_id: user.id,
-      email: user.email,
-      stripe_customer_id: customerId,
-    });
-  }
-
-  return customerId;
-};
 
 // Checkout session creation
 const createCheckoutSession = async (
